requirements.txt
opencv-python
numpy
pygame
matplotlib
pillow
keras
scikit-learn
tensorflow